pyproject.toml
[tool.poetry]
name = "user-management-api"
version = "0.1.0"
description = "REST API for user management using LiteStar"
authors = ["Your Name <your.email@example.com>"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.0"
litestar = {extras = ["standard"], version = "^2.5.1"}
uvicorn = "^0.27.0"
sqlalchemy = "^2.0.27"
psycopg = {extras = ["binary"], version = "^3.1.18"}
pydantic = "^2.6.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 